+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/calendar/CalendarPanel.java
@@ -78,6 +78,9 @@ public final class CalendarPanel extends JPanel {
 
   private static final ResourceBundle MESSAGES = ResourceBundle.getBundle(CalendarPanel.class.getName());
 
+  /**
+   * The default keyboard shortcuts.
+   */
   public static final KeyboardShortcuts<KeyboardShortcut> KEYBOARD_SHORTCUTS = keyboardShortcuts(KeyboardShortcut.class, new DefaultKeyboardShortcuts());
 
   /**
@@ -151,6 +154,7 @@ public final class CalendarPanel extends JPanel {
     formattedDateLabel = new JLabel("", SwingConstants.CENTER);
     formattedDateLabel.setBorder(emptyBorder());
     initializeUI();
+    addKeyEvents(builder.keyboardShortcuts);
     updateFormattedDate();
     bindEvents();
   }
@@ -266,12 +270,21 @@ public final class CalendarPanel extends JPanel {
     Builder includeTime(boolean includeTime);
 
     /**
+     * @param keyboardShortcut the keyboard shortcut key
+     * @param keyStroke the keyStroke to assign to the given shortcut key, null resets to the default one
+     * @return this builder instance
+     */
+    Builder keyStroke(KeyboardShortcut keyboardShortcut, KeyStroke keyStroke);
+
+    /**
      * @return a new {@link CalendarPanel} based on this builder
      */
     CalendarPanel build();
   }
 
   private static final class DefaultBuilder implements Builder {
+
+    private final KeyboardShortcuts<KeyboardShortcut> keyboardShortcuts = KEYBOARD_SHORTCUTS.copy();
 
     private LocalDateTime initialValue;
     private boolean includeTime = false;
@@ -291,6 +304,12 @@ public final class CalendarPanel extends JPanel {
     @Override
     public Builder includeTime(boolean includeTime) {
       this.includeTime = includeTime;
+      return this;
+    }
+
+    @Override
+    public Builder keyStroke(KeyboardShortcut keyboardShortcut, KeyStroke keyStroke) {
+      keyboardShortcuts.keyStroke(keyboardShortcut).set(keyStroke);
       return this;
     }
 
@@ -405,7 +424,6 @@ public final class CalendarPanel extends JPanel {
     setBorder(emptyBorder());
     add(createNorthPanel(), BorderLayout.NORTH);
     add(createDayPanel(), BorderLayout.CENTER);
-    addKeyEvents();
     layoutDayPanel();
   }
 
@@ -533,44 +551,44 @@ public final class CalendarPanel extends JPanel {
     formattedDateLabel.setText(dateFormatter.format(getLocalDateTime()) + (includeTime ? " " + timeFormatter.format(getLocalDateTime()) : ""));
   }
 
-  private void addKeyEvents() {
+  private void addKeyEvents(KeyboardShortcuts<KeyboardShortcut> keyboardShortcuts) {
     KeyEvents.Builder keyEvent = KeyEvents.builder()
             .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT);
-    keyEvent.keyStroke(KEYBOARD_SHORTCUTS.keyStroke(PREVIOUS_YEAR).get())
+    keyEvent.keyStroke(keyboardShortcuts.keyStroke(PREVIOUS_YEAR).get())
             .action(control(this::previousYear))
             .enable(this);
-    keyEvent.keyStroke(KEYBOARD_SHORTCUTS.keyStroke(NEXT_YEAR).get())
+    keyEvent.keyStroke(keyboardShortcuts.keyStroke(NEXT_YEAR).get())
             .action(control(this::nextYear))
             .enable(this);
-    keyEvent.keyStroke(KEYBOARD_SHORTCUTS.keyStroke(PREVIOUS_MONTH).get())
+    keyEvent.keyStroke(keyboardShortcuts.keyStroke(PREVIOUS_MONTH).get())
             .action(control(this::previousMonth))
             .enable(this);
-    keyEvent.keyStroke(KEYBOARD_SHORTCUTS.keyStroke(NEXT_MONTH).get())
+    keyEvent.keyStroke(keyboardShortcuts.keyStroke(NEXT_MONTH).get())
             .action(control(this::nextMonth))
             .enable(this);
-    keyEvent.keyStroke(KEYBOARD_SHORTCUTS.keyStroke(PREVIOUS_WEEK).get())
+    keyEvent.keyStroke(keyboardShortcuts.keyStroke(PREVIOUS_WEEK).get())
             .action(control(this::previousWeek))
             .enable(this);
-    keyEvent.keyStroke(KEYBOARD_SHORTCUTS.keyStroke(NEXT_WEEK).get())
+    keyEvent.keyStroke(keyboardShortcuts.keyStroke(NEXT_WEEK).get())
             .action(control(this::nextWeek))
             .enable(this);
-    keyEvent.keyStroke(KEYBOARD_SHORTCUTS.keyStroke(PREVIOUS_DAY).get())
+    keyEvent.keyStroke(keyboardShortcuts.keyStroke(PREVIOUS_DAY).get())
             .action(control(this::previousDay))
             .enable(this);
-    keyEvent.keyStroke(KEYBOARD_SHORTCUTS.keyStroke(NEXT_DAY).get())
+    keyEvent.keyStroke(keyboardShortcuts.keyStroke(NEXT_DAY).get())
             .action(control(this::nextDay))
             .enable(this);
     if (includeTime) {
-      keyEvent.keyStroke(KEYBOARD_SHORTCUTS.keyStroke(PREVIOUS_HOUR).get())
+      keyEvent.keyStroke(keyboardShortcuts.keyStroke(PREVIOUS_HOUR).get())
               .action(control(this::previousHour))
               .enable(this);
-      keyEvent.keyStroke(KEYBOARD_SHORTCUTS.keyStroke(NEXT_HOUR).get())
+      keyEvent.keyStroke(keyboardShortcuts.keyStroke(NEXT_HOUR).get())
               .action(control(this::nextHour))
               .enable(this);
-      keyEvent.keyStroke(KEYBOARD_SHORTCUTS.keyStroke(PREVIOUS_MINUTE).get())
+      keyEvent.keyStroke(keyboardShortcuts.keyStroke(PREVIOUS_MINUTE).get())
               .action(control(this::previousMinute))
               .enable(this);
-      keyEvent.keyStroke(KEYBOARD_SHORTCUTS.keyStroke(NEXT_MINUTE).get())
+      keyEvent.keyStroke(keyboardShortcuts.keyStroke(NEXT_MINUTE).get())
               .action(control(this::nextMinute))
               .enable(this);
     }